move the internal classes to another package
