+++ src/main/scala/nl/knaw/dans/easy/license/internal/DatasetValidator.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.license
+package nl.knaw.dans.easy.license.internal
 
 object DatasetValidator {
