+++ src/main/scala/nl/knaw/dans/easy/license/internal/Ldap.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.license
+package nl.knaw.dans.easy.license.internal
 
 import javax.naming.directory.{Attributes, SearchControls}
 import javax.naming.ldap.LdapContext
