+++ src/main/scala/nl/knaw/dans/easy/license/internal/PlaceholderMapper.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.license
+package nl.knaw.dans.easy.license.internal
 
 import java.io._
 import java.net.URLEncoder
@@ -21,7 +21,8 @@ import java.{util => ju}
 
 import nl.knaw.dans.common.lang.dataset.AccessCategory
 import nl.knaw.dans.common.lang.dataset.AccessCategory._
-import nl.knaw.dans.easy.license.FileAccessRight._
+import nl.knaw.dans.easy.license.BaseParameters
+import nl.knaw.dans.easy.license.internal.FileAccessRight._
 import nl.knaw.dans.pf.language.emd.types.{IsoDate, MetadataItem}
 import nl.knaw.dans.pf.language.emd.{EasyMetadata, EmdDate, Term}
 import org.apache.commons.codec.binary.Base64
